backend/pyproject.toml
[project]
name = "hisabkitab-backend"
version = "1.0.0"
description = "HisabKitab - Personal Expense Management Backend"
authors = [
    {name = "Niranjan Sah", email = "niranjansah250@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    "sqlalchemy>=2.0.36",
    "alembic==1.13.1",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.2",
    "pydantic>=2.9.1",
    "pydantic-settings>=2.5.2",
    "email-validator>=2.2.0",
    "google-generativeai==0.3.2",
    "openai-whisper>=20240930",
    "pytesseract==0.3.10",
    "Pillow>=11.0.0",
    "python-dateutil==2.8.2",
    "httpx==0.26.0",
    "aiofiles==23.2.1",
    "loguru==0.7.2",
    "firebase-admin==6.5.0",
    "resend==0.8.0",
    "boto3==1.34.34",
    "jinja2==3.1.3",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "slowapi>=0.1.9",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
